+++ sample/sample32-migrations/migrations/1481521933-second-release-changes.ts
@@ -4,11 +4,11 @@ import {QueryRunner} from "../../../src/query-runner/QueryRunner";
 
 export class SecondReleaseMigration1481521933 implements MigrationInterface {
 
-    async up(queryRunner: QueryRunner, connection: Connection): Promise<any> {
+    async up(queryRunner: QueryRunner): Promise<any> {
         await queryRunner.query("ALTER TABLE `post` CHANGE `name` `title` VARCHAR(500)");
     }
 
-    async down(queryRunner: QueryRunner, connection: Connection): Promise<any> {
+    async down(queryRunner: QueryRunner): Promise<any> {
         await queryRunner.query("ALTER TABLE `post` CHANGE `title` `name` VARCHAR(255)");
     }