+++ src/commands/MigrationCreateCommand.ts
@@ -64,10 +64,10 @@ export class MigrationCreateCommand {
 
 export class ${name}${timestamp} implements MigrationInterface {
 
-    public async up(queryRunner: QueryRunner, connection: Connection, entityManager?: EntityManager): Promise<any> {
+    public async up(queryRunner: QueryRunner): Promise<any> {
     }
 
-    public async down(queryRunner: QueryRunner, connection: Connection, entityManager?: EntityManager): Promise<any> {
+    public async down(queryRunner: QueryRunner): Promise<any> {
     }
 
 }